[brain tests] Burst dateutil from the main file
