+++ tests/brain/test_brain.py
@@ -24,13 +24,6 @@ from astroid.nodes.node_classes import Const
 from astroid.nodes.scoped_nodes import ClassDef
 
 try:
-    import dateutil  # pylint: disable=unused-import
-
-    HAS_DATEUTIL = True
-except ImportError:
-    HAS_DATEUTIL = False
-
-try:
     import attr as attr_module  # pylint: disable=unused-import
 
     HAS_ATTR = True
@@ -646,19 +639,6 @@ class EnumBrainTest(unittest.TestCase):
         for node in (attribute_nodes[1], name_nodes[1]):
             with pytest.raises(InferenceError):
                 node.inferred()
-
-
-@unittest.skipUnless(HAS_DATEUTIL, "This test requires the dateutil library.")
-class DateutilBrainTest(unittest.TestCase):
-    def test_parser(self):
-        module = builder.parse(
-            """
-        from dateutil.parser import parse
-        d = parse('2000-01-01')
-        """
-        )
-        d_type = next(module["d"].infer())
-        self.assertEqual(d_type.qname(), "datetime.datetime")
 
 
 class PytestBrainTest(unittest.TestCase):
